Fixed linux build!

+++ BFW/Headers/BFW_FileSystem.hpp
@@ -40,7 +40,7 @@ namespace BFW
 			const bool Save(std::ofstream& _File) const;
 			const bool Save(std::fstream& _File) const;
 
-			const size_t Hash() const;
+			const uint64_t Hash() const;
 
 			uint8_t* GetData();
 			const uint8_t* GetData() const;
